reverting file renaming
